adding a deletion alert whit sweetalert2

+++ client/src/components/TaskList.js
@@ -2,6 +2,7 @@ import React from 'react';
 import {useHistory,useParams} from 'react-router-dom';
 import Task from './Task';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 
 function TaskList({list,setList}) {
@@ -13,11 +14,26 @@ function TaskList({list,setList}) {
   }; 
 
   const deleteTask = id =>{
-    axios.delete(`/api/tasks/delete/${id}`)
+    Swal.fire({
+      title:'Alert',
+      text:'Are you sure you want to delete this task?',
+      icon:'warning',
+      showCancelButton:true
+    }).then(result=>{
+      if(result.value){
+        axios.delete(`/api/tasks/delete/${id}`)
           .then(res=>{
             const newTasksList = list.filter((actualTasks) => actualTasks._id !== id);
             setList(newTasksList);
           })
+          .catch(err => Swal.fire({
+            icon:'error',
+            title:'Error while deleting',
+            text:'An error occurred while deleting'
+          }))
+      }
+    })
+   
   }
 
   return (
